Updates for Emergency_situation
